% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mslca.R
\name{count_mslca}
\alias{count_mslca}
\title{LCA analysis}
\usage{
count_mslca(data_input, lcaname, ms)
}
\arguments{
\item{data_input}{data.frame obtained from read.csv}

\item{lcaname}{name of the taxonomy level, for example "kingdom"}

\item{ms}{MS.count as input}
}
\description{
perform aggregation on specified lca level with the count of MS2
}
